+++ src/app/backend/abstracts/manga-series.tsx
@@ -1,0 +1,5 @@
+interface MangaSeries {
+
+}
+
+export default MangaSeries