+++ src/app/backend/search/search-result.tsx
@@ -1,0 +1,5 @@
+interface SearchResult {
+
+}
+
+export default SearchResult